Consolidate internal Common and Compat modules into a single module named Internal
