+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Bifoldable.hs
@@ -23,7 +23,7 @@ import Data.Functor.Classes (Eq2,Show2)
 import Test.QuickCheck.Property (Property)
 import Data.Monoid
 import Data.Orphans ()
-import Test.QuickCheck.Classes.Common
+import Test.QuickCheck.Classes.Internal
 #endif
 
 #if HAVE_BINARY_LAWS
