+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Semigroup.hs
@@ -17,7 +17,7 @@ import Data.Proxy (Proxy)
 import Test.QuickCheck hiding ((.&.))
 import Test.QuickCheck.Property (Property)
 
-import Test.QuickCheck.Classes.Common (Laws(..), SmallList(..), myForAllShrink)
+import Test.QuickCheck.Classes.Internal (Laws(..), SmallList(..), myForAllShrink)
 
 import Data.Foldable (foldr1,toList)
 import Data.List.NonEmpty (NonEmpty((:|)))
